requirements.txt
streamlit
pandas
pillow
openai
openpyxl
pillow-heif
